% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_calculator_functions.R
\docType{data}
\name{data.taxrebate}
\alias{data.taxrebate}
\alias{data.taxrebate_under87A}
\title{tax rebates}
\format{
data.taxrebate: A data frame (11, 6):
\describe{
\item{rebate.id}{rebate identification number}
\item{rebate.code}{rebate code/short name}
\item{rebate.limit}{maximum rebate}
\item{tax_regime}{old or new regime}
\item{financial_year}{financial year}
\item{rebate.info}{comments or additional information}
}

data.taxrebate: A data frame (11, 6):
\describe{
\item{rebate.id}{rebate identification number}
\item{rebate.code}{rebate code/short name}
\item{rebate.limit}{maximum rebate}
\item{tax_regime}{old or new regime}
\item{financial_year}{financial year}
\item{rebate.info}{comments or additional information}
}
}
\usage{
data.taxrebate

data.taxrebate_under87A
}
\description{
tax rebates before tax slab adjustments

tax rebates before tax slab adjustments
}
\keyword{datasets}
