% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.taxslabs}
\alias{data.taxslabs}
\title{tax slabs for both old and new regime}
\format{
data.taxslabs: A data frame (10, 7):
\describe{
  \item{slab.id}{slab identification number}
  \item{slab.min}{slab minimum amount}
  \item{slab.max}{slab maximum amount}
  \item{slab.percent}{tax percentage charged for the slab}
  \item{tax_regime}{old or new regime}
  \item{financial_year}{financial year}
  \item{slab.info}{comments or additional information}
}
}
\usage{
data.taxslabs
}
\description{
tax slabs for both old and new regime
}
\keyword{datasets}
