% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_calculator_functions.R
\name{print.ITR}
\alias{print.ITR}
\title{print.ITR}
\usage{
\method{print}{ITR}(x, partial = TRUE, ...)
}
\arguments{
\item{x}{a ITR class object}

\item{partial}{logical, whether to print only a subset of columns}

\item{...}{additional parameters for print.data.frame}
}
\description{
print method for class "ITR"
}
